requirements.txt
PyYAML>=6.0
python-dotenv
openai
feedparser>=6.0.10
streamlit
pandas
trafilatura>=1.6.0
requests>=2.31.0
python-dateutil>=2.8.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
beautifulsoup4>=4.11.0
